Add basic episode support, halp pls
Something weird is happening with the loading of these episodes, i've
added like 1000 logs and everything seems fine and I've tried heaps of
things but episodes seem to load fine for the first serie you select but
selecting another serie and viewing episodes has a high change of never
loading. Cause of this seems to be the cache clearering i've done
because the cache needs to be cleared otherwise when you select other
episodes there is a short delay where it removes the old episodes (and
torrent controls).
I don't see what the problem is.