Support for user defined AutoDownload Settings #360
Still needs work and better code. I couldn't figure out a way to easily
make it so you can't enter characters other than numbers but I just
added an error and disabled the save button if that happens. Also
couldn't figure out a way to set the input to the current saved value
but make it so when you change it it doesn't update the little box that
displays the current setting so I just created 2 variables for each, the
saved one and a custom one, not ideal.