Sweet rewrite of the calendar event hook to prevent the nasty calendar refresh mechanism.
Now the CalendarEvents service receives the calendar's start and end date and fetches episodes for that (also when you change to week mode or change the date)
Also the favoritesservice passes all the new episodes for a new show to the calendar after adding it, upon which the calendar checks if there's episodes airing on one of the registered days and adds it to the existing array.

When the calendar changes, the old dates are removed from cache and new ones are populated.
When a show is removed, it hides the existing episodes from the calendar and they toggle nicely when switching the 'show on calendar' and specials flags.