Fixed annoying tooltip positions in Episode List and optimised using CSS
Fixed the stupid tooltips loading where ever they please. Manually
aligned some and just manually set the width so that they could fit and
had to change the tooltip around for the watched button because it's
right on the edge. Tested most languages, seems to work fine.

Also swapped out the bad on mouseenter/exit events to show the torrent
controls for each episode and switched it to a much more efficient CSS
based show/hide also fixes some of the ghosting that happens when you
hovered and went onto a tooltip which doesnt register the mouseleave
action.