Update vendors
Updated angular 1.4.3 and bootstrap to a 0.13.1, says it's a snapshot
but it's the recommended version. Also removed a bunch of modules that
we didn't need which dropped the file size considerably and I've noticed
some performance increases.

As always, be on the lookout for any errors these upgrades could cause.