Improve backgroundRotator transitions and added cooldown
Added cooldown to prevent it from being spammed although this is mostly
redundant now because I moved the background:load event further down the
FavoritesService chain so it only broadcasts once it has been added as
opposed to immediately which fired all series adding at once.

Changed the way it works as well using a class to show and hide, in my
tests this fixes 'stuttering' from happening which occurs when directly
altering the style of the element.