Hopefully improved calendar overflowing and initial episodesGrid concept for ToDo view
Made some changes to when the scrollbar is shown using CSS. When
seriesList is active it will always be hidden as there should be no case
where the seriesList would be open and you can view calendar and by
default, overflow is disabled while sidePanel open unless the
scrollheight is greater than the window height.

And for the ToDo grid, just added a basic function that centers the
episodes, it's not exactly ideal but was just a quick mockup.