General Tooltip fixes and calendar event touchups
Compromised and set a max-width on the calendar tooltips to the left
most tooltips won't overflow, this is the simplest solution and still
works fine I think.
Fixed 1px gap between tooltip-inner and tooltip-arrow
Fixed tooltips lineheight causing it quickly change size when it loads.
Fixed text-decorations on the calendar events when hovering when they
were already 'line-through' and made it so the episode numbers correctly
underlined when hovered.
Made the events slightly larger too (.6px to be exact) to fix the
paddings displaying incorrectly when an event was marked as watched