fix torrentz seeders and leechers by removing the comma
this is an old bug I just discovered.
Torrentz presents seeders and leeches as strings of numbers with commas. eg. 2,199 instead of 2199.
this leads the auto-download to perform unexpectedly in its seeders evaluation, because it uses 
parseInt(items[0].seeders, 10) >= minSeeders
to determine if there are sufficient seeders to launch magnet.
the problem? parseInt("2,199",10) returns 2 not 2199 !
so torrentz users have been missing out on autodownloaded magnets whenever seeders get higher than 999 :-)