W.I.P. for #668
torrent dialogue sort-columns feature.

The standard torrent dialog columns: Size, Seeders and Leechers can now
be sorted in descending order.

Unfortunately, it is not practical to add a column for Age, because all
the search providers present this information in different ways.

However, since sort-by-age was the primary reason for raising this
feature, I have elected to provide a setting, in the
torrent-search-engines panel, to enable a Sort Menu on the torrent
dialog.
The default is for the sort menu to be hidden.
The Sort menu includes an entry to sort the search results by Age, in
newest to oldest order.