testing for #712
the window.onload is redundant. the .run is always invoked at start-up
and does not need the additional delay of the onload event being
triggered.
While previously the onload event in this code did not interfere with
setting up the midnight timer and backup timer, now the event is not
triggered at start-up anymore probably because the page is loading
faster.

Also found that the settimer with a negative ms value is not longer
triggering. changes to use 0 instead.