change the default settings for the TorrentDialog to true.
The natural inclinition of users, is to assume that when the used the
DuckieTV Torrent Search Engine settings, that this would apply to
searches they make using the TorrentDialog.
By Defaulting the Advanced Options of the TorrentDialog check boxes to
true, we support that natural assumption.
And by Defaulting the Show Advanced Options to true, the uesrs can see
that the have a choice on the TorrentDialog to turn off settings if they
want to, and even hide the Advanced Settings if they choose to.