delay infoHash fetch to allow host time to add torrent
sometimes our torrent clients are doing a torrent infohash cleanup too
soon after submitting a torrent to the host (within 0.3s), and we end up
removing the infohash from the db before the torrent host has had time
to set the torrent up.
this will delay the clean up by 5 seconds, which should be more than
enough time for the host to add the torrent and respond to infohash
request from our client.