Make Standalone shutdown completely.
Since the activation of the background task, I've noticed that the
Standalone version does not always completely shutdown when using either
the X button on the main window or the Exit menu from the systray.
The main windows appears to close (or is hidden?), but the task manager
shows four nwjs processes that continue to run forever.
Using App.quit() ensures complete shutdown.
However, this may also interupt background i/o to DB, so is there a
better solution?